requirements.txt
APScheduler==3.11.0
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
speedtest-cli==2.1.3
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzlocal==5.3.1
Werkzeug==3.1.3
